requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cssselect2==0.8.0
decorator==5.2.1
distlib==0.3.9
Django==5.1.2
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-extensions==4.1
django-js-asset==3.1.2
django-timezone-field==7.1
djangorestframework==3.15.2
executing==2.2.0
filelock==3.17.0
fonttools==4.58.4
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
reportlab==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.1
stack-data==0.6.3
stripe==12.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
weasyprint==65.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zopfli==0.2.3.post1
